downloader/pyproject.toml
[tool.poetry]
name = "bitbucket_repo_downloader"
version = "0.1.0"
description = "Download all repositories from the bitbucket account"
authors = ["Kirill Chertkov <chertkov.k@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.22.0"
gitpython = "^3.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
